pyproject.toml
[tool.poetry]
name = "drweb"
version = "0.1.0"
description = ""
authors = ["Igor"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.2"
sqlalchemy = "^2.0.32"
asyncio = "^3.4.3"
uvicorn = "^0.30.6"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
pydantic-settings = "^2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
